Split GUI app wiring into focused modules
